feat (budget): Begins adding budget creation section.
