+++ Breeze.NewWeb/src/config/routing/AppRoutes.tsx
@@ -3,7 +3,8 @@ import { LandingPage } from '../../pages/LandingPage'
 import { BudgetProvider } from '../../services/providers/BudgetProvider'
 import { useMsal } from '@azure/msal-react'
 import { Navigation } from '../../components/navigation/Navigation'
-import { BudgetCarousel } from '../../components/dashboard/BudgetCarousel'
+import { BudgetCarousel } from '../../pages/dashboard/BudgetCarousel'
+import { CreateBudget } from '../../pages/budget/CreateBudget'
 
 export const AppRoutes = () => {
 	const { accounts } = useMsal()
@@ -12,6 +13,15 @@ export const AppRoutes = () => {
 	return (
 		<BrowserRouter>
 			<Routes>
+				<Route
+					path='/login'
+					element={
+						<>
+							<Navigation />
+							<LandingPage />
+						</>
+					}
+				/>
 				<Route
 					path='/'
 					element={
@@ -28,12 +38,18 @@ export const AppRoutes = () => {
 					}
 				/>
 				<Route
-					path='/login'
+					path='/create-budget'
 					element={
-						<>
-							<Navigation />
-							<LandingPage />
-						</>
+						isAuthenticated ? (
+							<BudgetProvider>
+								<div className='h-screen w-screen flex flex-col justify-center items-center'>
+									<Navigation />
+									<CreateBudget />
+								</div>
+							</BudgetProvider>
+						) : (
+							<Navigate to='/login' />
+						)
 					}
 				/>
 				<Route
